requirements.txt
fsspec==2025.2.0
adlfs==2024.12.0
polars==1.24.0
pandas==2.0.0
deltalake==0.25.4
python-dotenv==1.0.1
numpy==1.22.4